Improve responsive hook
