+++ packages/canon/src/components/Button/index.tsx
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 export { Button } from './Button';
-export type { ButtonProps } from './Button';
+export type { ButtonProps } from './types';
