+++ packages/canon/src/components/Heading/Heading.tsx
@@ -17,14 +17,16 @@
 import React, { forwardRef } from 'react';
 import { HeadingProps } from './types';
 import { useTheme } from '../../theme/context';
-import { getResponsiveValue } from '../../utils/getResponsiveValue';
 
 export const Heading = forwardRef<HTMLHeadingElement, HeadingProps>(
   (props, ref) => {
     const { children, variant = 'title1', as = 'h1', ...restProps } = props;
-    const { breakpoint } = useTheme();
-    const responsiveVariant = getResponsiveValue(variant, breakpoint);
+    const { getResponsiveValue } = useTheme();
 
+    // Get the responsive value for the variant
+    const responsiveVariant = getResponsiveValue(variant);
+
+    // Determine the component to render based on the variant
     let Component = as;
     if (variant === 'title2') Component = 'h2';
     if (variant === 'title3') Component = 'h3';
